+++ backend/indexing/indexingFiles.go
@@ -10,6 +10,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/gtsteffaniak/filebrowser/backend/common/errors"
 	"github.com/gtsteffaniak/filebrowser/backend/common/settings"
 	"github.com/gtsteffaniak/filebrowser/backend/common/utils"
 	"github.com/gtsteffaniak/filebrowser/backend/indexing/iteminfo"
@@ -39,6 +40,7 @@ type Index struct {
 	CurrentSchedule            int `json:"-"`
 	settings.Source            `json:"-"`
 	Directories                map[string]*iteminfo.FileInfo `json:"-"`
+	DirectoriesLedger          map[string]bool               `json:"-"`
 	runningScannerCount        int                           `json:"-"`
 	SmartModifier              time.Duration                 `json:"-"`
 	FilesChangedDuringIndexing bool                          `json:"-"`
@@ -66,9 +68,10 @@ func init() {
 func Initialize(source settings.Source, mock bool) {
 	indexesMutex.Lock()
 	newIndex := Index{
-		mock:        mock,
-		Source:      source,
-		Directories: make(map[string]*iteminfo.FileInfo),
+		mock:              mock,
+		Source:            source,
+		Directories:       make(map[string]*iteminfo.FileInfo),
+		DirectoriesLedger: make(map[string]bool),
 	}
 	newIndex.ReducedIndex = ReducedIndex{
 		Status:     "indexing",
@@ -83,7 +86,7 @@ func Initialize(source settings.Source, mock bool) {
 		newIndex.RunIndexing("/", false)
 		go newIndex.setupIndexingScanners()
 	} else {
-		newIndex.Status = "disabled"
+		newIndex.Status = "ready"
 		logger.Debug("indexing disabled for source: " + newIndex.Name)
 	}
 }
@@ -94,7 +97,7 @@ func (idx *Index) indexDirectory(adjustedPath string, quick, recursive bool) err
 	// Open the directory
 	dir, err := os.Open(realPath)
 	if err != nil {
-		idx.RemoveDirectory(adjustedPath) // Remove, must have been deleted
+		// must have been deleted
 		return err
 	}
 	defer dir.Close()
@@ -103,6 +106,20 @@ func (idx *Index) indexDirectory(adjustedPath string, quick, recursive bool) err
 	if err != nil {
 		return err
 	}
+
+	// check if excluded from indexing
+	hidden := isHidden(dirInfo, idx.Path+adjustedPath)
+	if !recursive && idx.shouldSkip(true, hidden, adjustedPath) {
+		return errors.ErrNotIndexed
+	}
+
+	// if indexing, mark the directory as valid and indexed.
+	if recursive {
+		// Prevent race conditions if scanning becomes concurrent in the future.
+		idx.mu.Lock()
+		idx.DirectoriesLedger[adjustedPath] = true
+		idx.mu.Unlock()
+	}
 	combinedPath := adjustedPath + "/"
 	if adjustedPath == "/" {
 		combinedPath = "/"
@@ -110,9 +127,9 @@ func (idx *Index) indexDirectory(adjustedPath string, quick, recursive bool) err
 	// get whats currently in cache
 	idx.mu.RLock()
 	cacheDirItems := []iteminfo.ItemInfo{}
-	modChange := true // default to true
+	modChange := false
 	cachedDir, exists := idx.Directories[adjustedPath]
-	if exists && quick {
+	if exists {
 		modChange = dirInfo.ModTime() != cachedDir.ModTime
 		cacheDirItems = cachedDir.Folders
 	}
@@ -120,48 +137,81 @@ func (idx *Index) indexDirectory(adjustedPath string, quick, recursive bool) err
 
 	// If the directory has not been modified since the last index, skip expensive readdir
 	// recursively check cached dirs for mod time changes as well
-	if !modChange && recursive {
-		for _, item := range cacheDirItems {
-			err = idx.indexDirectory(combinedPath+item.Name, quick, true)
-			if err != nil {
-				logger.Errorf("error indexing directory %v : %v", combinedPath+item.Name, err)
+	if recursive {
+		if modChange {
+			idx.mu.Lock()
+			idx.FilesChangedDuringIndexing = true
+			idx.mu.Unlock()
+		} else if quick {
+			for _, item := range cacheDirItems {
+				err = idx.indexDirectory(combinedPath+item.Name, quick, true)
+				if err != nil && err != errors.ErrNotIndexed {
+					logger.Errorf("error indexing directory %v : %v", combinedPath+item.Name, err)
+				}
 			}
+			return nil
 		}
-		return nil
 	}
-
-	if quick {
-		idx.mu.Lock()
-		idx.FilesChangedDuringIndexing = true
-		idx.mu.Unlock()
+	dirFileInfo, err2 := idx.GetDirInfo(dir, dirInfo, realPath, adjustedPath, combinedPath, quick, recursive)
+	if err2 != nil {
+		return err2
 	}
+	// Update the current directory metadata in the index
+	idx.UpdateMetadata(dirFileInfo)
+	return nil
+}
 
-	// Read directory contents
-	files, err := dir.Readdir(-1)
+func (idx *Index) GetFsDirInfo(adjustedPath string) (*iteminfo.FileInfo, error) {
+	realPath, isDir, err := idx.GetRealPath(adjustedPath)
 	if err != nil {
-		return err
+		return nil, err
 	}
+	if !isDir {
+		return nil, fmt.Errorf("path is not a directory: %s", adjustedPath)
+	}
+	dir, err := os.Open(realPath)
+	if err != nil {
+		return nil, err
+	}
+	defer dir.Close()
 
+	dirInfo, err := dir.Stat()
+	if err != nil {
+		return nil, err
+	}
+	fmt.Println(dir.Name())
+	combinedPath := adjustedPath + "/"
+	if adjustedPath == "/" {
+		combinedPath = "/"
+	}
+	return idx.GetDirInfo(dir, dirInfo, realPath, adjustedPath, combinedPath, false, false)
+}
+
+func (idx *Index) GetDirInfo(dirInfo *os.File, stat os.FileInfo, realPath, adjustedPath, combinedPath string, quick, recursive bool) (*iteminfo.FileInfo, error) {
+	// Read directory contents
+	files, err := dirInfo.Readdir(-1)
+	if err != nil {
+		return nil, err
+	}
 	var totalSize int64
 	fileInfos := []iteminfo.ItemInfo{}
 	dirInfos := []iteminfo.ItemInfo{}
 
 	// Process each file and directory in the current directory
 	for _, file := range files {
-		isHidden := isHidden(file, idx.Path+combinedPath)
+		hidden := isHidden(file, idx.Path+combinedPath)
 		isDir := iteminfo.IsDirectory(file)
 		fullCombined := combinedPath + file.Name()
-		if idx.shouldSkip(isDir, isHidden, fullCombined) {
+		if idx.shouldSkip(isDir, hidden, fullCombined) {
 			continue
 		}
 		itemInfo := &iteminfo.ItemInfo{
 			Name:    file.Name(),
 			ModTime: file.ModTime(),
-			Hidden:  isHidden,
+			Hidden:  hidden,
 		}
 
 		if isDir {
-
 			// skip non-indexable dirs.
 			if file.Name() == "$RECYCLE.BIN" || file.Name() == "System Volume Information" {
 				continue
@@ -169,6 +219,8 @@ func (idx *Index) indexDirectory(adjustedPath string, quick, recursive bool) err
 
 			dirPath := combinedPath + file.Name()
 			if recursive {
+				// clear for garbage collection
+				file = nil
 				// Recursively index the subdirectory
 				err = idx.indexDirectory(dirPath, quick, recursive)
 				if err != nil {
@@ -192,9 +244,11 @@ func (idx *Index) indexDirectory(adjustedPath string, quick, recursive bool) err
 			idx.NumFiles++
 		}
 	}
+
 	if totalSize == 0 && idx.Config.IgnoreZeroSizeFolders {
-		return nil
+		return nil, errors.ErrNotIndexed
 	}
+
 	// Create FileInfo for the current directory
 	dirFileInfo := &iteminfo.FileInfo{
 		Path:    adjustedPath,
@@ -202,16 +256,13 @@ func (idx *Index) indexDirectory(adjustedPath string, quick, recursive bool) err
 		Folders: dirInfos,
 	}
 	dirFileInfo.ItemInfo = iteminfo.ItemInfo{
-		Name:    dirInfo.Name(),
+		Name:    filepath.Base(dirInfo.Name()),
 		Type:    "directory",
 		Size:    totalSize,
-		ModTime: dirInfo.ModTime(),
+		ModTime: stat.ModTime(),
 	}
-
 	dirFileInfo.SortItems()
-	// Update the current directory metadata in the index
-	idx.UpdateMetadata(dirFileInfo)
-	return nil
+	return dirFileInfo, nil
 }
 
 // input should be non-index path.
@@ -280,7 +331,7 @@ func (idx *Index) RefreshFileInfo(opts iteminfo.FileOptions) error {
 	}
 	err := idx.indexDirectory(refreshOptions.Path, false, false)
 	if err != nil {
-		return fmt.Errorf("file/folder does not exist to refresh data: %s", refreshOptions.Path)
+		return err
 	}
 	file, exists := idx.GetMetadataInfo(refreshOptions.Path, true)
 	if !exists {